Reorganize world module
