+++ src/world/mod.rs
@@ -1,33 +1,26 @@
+mod level;
 pub mod level_settings;
-
-use std::sync::LazyLock;
 
 use avian2d::prelude::{Collider, Gravity, RigidBody};
 use bevy::{
     app::{plugin_group, Plugin, Startup, Update},
     asset::AssetServer,
     math::Vec2,
-    prelude::{
-        Bundle, Commands, Component, Event, EventReader, EventWriter, IntoSystemConfigs, Res,
-        ResMut,
-    },
+    prelude::{Bundle, Commands, Component, IntoSystemConfigs, Res},
     utils::default,
 };
 use bevy_ecs_ldtk::{
     app::{LdtkIntCell, LdtkIntCellAppExt},
-    LdtkWorldBundle, LevelSelection,
+    LdtkWorldBundle,
 };
+use level::{change_level, initial_level_change, ChangeLevel};
 use level_settings::{update_level_settings, LevelSettings};
 
-use crate::{impl_intcell, player::SyncCameraWithPlayer};
+use crate::impl_intcell;
 
 static WORLD_PATH: &'static str = "world.ldtk";
 
 pub const GRAVITY: Gravity = Gravity(Vec2::new(0., -98.1));
-
-// TODO: remove
-static FIRST_LEVEL: LazyLock<LevelSelection> =
-    LazyLock::new(|| LevelSelection::iid("977043b0-c210-11ef-82b1-d58aa0d63a63"));
 
 plugin_group! {
     pub struct WorldPlugins {
@@ -49,32 +42,11 @@ impl Plugin for BasePlugin {
     }
 }
 
-fn initial_level_change(mut commands: Commands) {
-    commands.insert_resource(FIRST_LEVEL.clone());
-    commands.send_event(ChangeLevel(FIRST_LEVEL.clone()));
-}
-
 fn spawn_world(mut commands: Commands, asset_server: Res<AssetServer>) {
     commands.spawn(LdtkWorldBundle {
         ldtk_handle: asset_server.load(WORLD_PATH).into(),
         ..default()
     });
-}
-
-#[derive(Event)]
-pub struct ChangeLevel(pub LevelSelection);
-
-fn change_level(
-    mut event_reader: EventReader<ChangeLevel>,
-    mut level_selection: ResMut<LevelSelection>,
-    mut event_writer: EventWriter<SyncCameraWithPlayer>,
-) {
-    let Some(last) = event_reader.read().last() else {
-        return;
-    };
-
-    *level_selection = last.0.clone();
-    event_writer.send_default();
 }
 
 pub trait IntCell {
